+++ e2e/node/basic/queryExpiry.test.ts
@@ -1,15 +1,16 @@
 import { beforeEach, describe, it, vi, expect } from 'vitest';
 import {
   MockReplica,
+  mockSyncTimeResponse,
   prepareV2QueryResponse,
   prepareV2ReadStateSubnetResponse,
-  prepareV2ReadStateTimeResponse,
 } from '../utils/mock-replica.ts';
 import { IDL } from '@icp-sdk/core/candid';
 import { Principal } from '@icp-sdk/core/principal';
-import { KeyPair, randomIdentity, randomKeyPair } from '../utils/identity.ts';
+import { randomIdentity, randomKeyPair } from '../utils/identity.ts';
 import {
   CertificateOutdatedErrorCode,
+  CertificateTimeErrorCode,
   HttpAgent,
   requestIdOf,
   TrustError,
@@ -84,6 +85,12 @@ describe('queryExpiry', () => {
   });
 
   it('should fail if the timestamp is outside the max ingress expiry (no retry)', async () => {
+    const timeDiffMsecs = 6 * MINUTE_TO_MSECS;
+    const futureDate = new Date(now.getTime() + timeDiffMsecs);
+
+    // advance to go over the max ingress expiry (5 minutes)
+    advanceTimeByMilliseconds(timeDiffMsecs);
+
     const agent = await HttpAgent.create({
       host: mockReplica.address,
       rootKey: subnetKeyPair.publicKeyDer,
@@ -105,21 +112,17 @@ describe('queryExpiry', () => {
     mockReplica.setV2QuerySpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(responseBody);
     });
-
-    // advance to go over the max ingress expiry (5 minutes)
-    advanceTimeByMilliseconds(6 * MINUTE_TO_MSECS);
-
     const { responseBody: subnetResponseBody } = await prepareV2ReadStateSubnetResponse({
       nodeIdentity,
       canisterRanges: [[canisterId.toUint8Array(), canisterId.toUint8Array()]],
       keyPair: subnetKeyPair,
-      date: now,
+      date: futureDate, // make sure the certificate is fresh for this call
     });
     mockReplica.setV2ReadStateSpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(subnetResponseBody);
     });
 
-    expect.assertions(5);
+    expect.assertions(4);
 
     try {
       await actor[greetMethodName](greetReq);
@@ -128,10 +131,12 @@ describe('queryExpiry', () => {
     }
 
     expect(mockReplica.getV2QuerySpy(canisterId.toString())).toHaveBeenCalledTimes(1);
-    expect(mockReplica.getV2ReadStateSpy(canisterId.toString())).toHaveBeenCalledTimes(0);
   });
 
   it('should retry and fail if the timestamp is outside the max ingress expiry (with retry)', async () => {
+    const timeDiffMsecs = 6 * MINUTE_TO_MSECS;
+    const futureDate = new Date(now.getTime() + timeDiffMsecs);
+
     const agent = await HttpAgent.create({
       host: mockReplica.address,
       rootKey: subnetKeyPair.publicKeyDer,
@@ -153,18 +158,9 @@ describe('queryExpiry', () => {
     mockReplica.setV2QuerySpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(responseBody);
     });
-    mockReplica.setV2QuerySpyImplOnce(canisterId.toString(), (_req, res) => {
-      res.status(200).send(responseBody);
-    });
-    mockReplica.setV2QuerySpyImplOnce(canisterId.toString(), (_req, res) => {
-      res.status(200).send(responseBody);
-    });
-    mockReplica.setV2QuerySpyImplOnce(canisterId.toString(), (_req, res) => {
-      res.status(200).send(responseBody);
-    });
 
     // advance to go over the max ingress expiry (5 minutes)
-    advanceTimeByMilliseconds(6 * MINUTE_TO_MSECS);
+    advanceTimeByMilliseconds(timeDiffMsecs);
 
     const { responseBody: subnetResponseBody } = await prepareV2ReadStateSubnetResponse({
       nodeIdentity,
@@ -172,8 +168,16 @@ describe('queryExpiry', () => {
       keyPair: subnetKeyPair,
       date: now,
     });
+    // fetch subnet keys, fails for certificate freshness checks
     mockReplica.setV2ReadStateSpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(subnetResponseBody);
+    });
+    // sync time, keeping a date in the future to make sure the agent still has outdated time
+    await mockSyncTimeResponse({
+      mockReplica,
+      keyPair: subnetKeyPair,
+      date: futureDate,
+      canisterId,
     });
 
     expect.assertions(5);
@@ -181,11 +185,14 @@ describe('queryExpiry', () => {
     try {
       await actor[greetMethodName](greetReq);
     } catch (e) {
-      expectCertificateOutdatedError(e);
+      expect(e).toBeInstanceOf(TrustError);
+      const err = e as TrustError;
+      expect(err.cause.code).toBeInstanceOf(CertificateTimeErrorCode);
+      expect(err.message).toContain('Certificate is signed more than 5 minutes in the past.');
     }
 
-    expect(mockReplica.getV2QuerySpy(canisterId.toString())).toHaveBeenCalledTimes(4);
-    expect(mockReplica.getV2ReadStateSpy(canisterId.toString())).toHaveBeenCalledTimes(1);
+    expect(mockReplica.getV2QuerySpy(canisterId.toString())).toHaveBeenCalledTimes(1);
+    expect(mockReplica.getV2ReadStateSpy(canisterId.toString())).toHaveBeenCalledTimes(4);
   });
 
   it('should not retry if the timestamp is outside the max ingress expiry (verifyQuerySignatures=false)', async () => {
@@ -233,7 +240,12 @@ describe('queryExpiry', () => {
     'should account for local clock drift (more than 5 minutes in the %s)',
     async (_, timeDiffMsecs) => {
       const replicaDate = new Date(now.getTime() + timeDiffMsecs);
-      await mockSyncTimeResponse({ mockReplica, keyPair: subnetKeyPair, date: replicaDate });
+      await mockSyncTimeResponse({
+        mockReplica,
+        keyPair: subnetKeyPair,
+        date: replicaDate,
+        canisterId: ICP_LEDGER,
+      });
 
       const agent = await HttpAgent.create({
         host: mockReplica.address,
@@ -308,16 +320,6 @@ describe('queryExpiry', () => {
       res.status(200).send(responseBody);
     });
 
-    const { responseBody: subnetResponseBody } = await prepareV2ReadStateSubnetResponse({
-      nodeIdentity,
-      canisterRanges: [[canisterId.toUint8Array(), canisterId.toUint8Array()]],
-      keyPair: subnetKeyPair,
-      date: replicaDate,
-    });
-    mockReplica.setV2ReadStateSpyImplOnce(canisterId.toString(), (_req, res) => {
-      res.status(200).send(subnetResponseBody);
-    });
-
     expect.assertions(4);
 
     try {
@@ -366,6 +368,12 @@ describe('queryExpiry', () => {
     mockReplica.setV2ReadStateSpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(subnetResponseBody);
     });
+    await mockSyncTimeResponse({
+      mockReplica,
+      keyPair: subnetKeyPair,
+      date: replicaDate,
+      canisterId,
+    });
     mockReplica.setV2ReadStateSpyImplOnce(canisterId.toString(), (_req, res) => {
       res.status(200).send(subnetResponseBody);
     });
@@ -374,7 +382,7 @@ describe('queryExpiry', () => {
 
     expect(actorResponse).toEqual(greetRes);
     expect(mockReplica.getV2QuerySpy(canisterId.toString())).toHaveBeenCalledTimes(1);
-    expect(mockReplica.getV2ReadStateSpy(canisterId.toString())).toHaveBeenCalledTimes(1);
+    expect(mockReplica.getV2ReadStateSpy(canisterId.toString())).toHaveBeenCalledTimes(4);
 
     const req = mockReplica.getV2QueryReq(canisterId.toString(), 0);
     expect(requestIdOf(req.content)).toEqual(requestId);
@@ -391,31 +399,4 @@ function expectCertificateOutdatedError(e: unknown) {
   const err = e as TrustError;
   expect(err.cause.code).toBeInstanceOf(CertificateOutdatedErrorCode);
   expect(err.message).toContain('Certificate is stale');
-}
-
-async function mockSyncTimeResponse({
-  mockReplica,
-  keyPair,
-  date,
-  canisterId,
-}: {
-  mockReplica: MockReplica;
-  keyPair: KeyPair;
-  date?: Date;
-  canisterId?: Principal | string;
-}) {
-  canisterId = Principal.from(canisterId ?? ICP_LEDGER).toText();
-  const { responseBody: timeResponseBody } = await prepareV2ReadStateTimeResponse({
-    keyPair,
-    date,
-  });
-  mockReplica.setV2ReadStateSpyImplOnce(canisterId, (_req, res) => {
-    res.status(200).send(timeResponseBody);
-  });
-  mockReplica.setV2ReadStateSpyImplOnce(canisterId, (_req, res) => {
-    res.status(200).send(timeResponseBody);
-  });
-  mockReplica.setV2ReadStateSpyImplOnce(canisterId, (_req, res) => {
-    res.status(200).send(timeResponseBody);
-  });
 }