+++ packages/agent/src/agent/http/index.ts
@@ -1239,11 +1239,11 @@ export class HttpAgent implements Agent {
             Array(3)
               .fill(null)
               .map(async () => {
-                // TODO: disable certificate freshness check for this request
                 const status = await canisterStatusRequest({
                   canisterId,
                   agent: anonymousAgent,
                   paths: ['time'],
+                  disableCertificateTimeVerification: true, // avoid recursive calls to syncTime
                 });
 
                 const date = status.get('time');