+++ packages/agent/src/canisterStatus/index.ts
@@ -13,12 +13,12 @@ import {
   HashTreeDecodeErrorCode,
   UnexpectedErrorCode,
   InputError,
+  CertificateTimeErrorCode,
 } from '../errors.ts';
 import { HttpAgent } from '../agent/http/index.ts';
 import {
   type Cert,
   Certificate,
-  type CreateCertificateOptions,
   flatten_forks,
   check_canister_ranges,
   LookupPathStatus,
@@ -116,20 +116,32 @@ export type Path =
 export type StatusMap = Map<Path | string, Status>;
 
 export type CanisterStatusOptions = {
+  /**
+   * The effective canister ID to use in the underlying {@link HttpAgent.readState} call.
+   */
   canisterId: Principal;
+  /**
+   * The agent to use to make the canister request. Must be authenticated.
+   */
   agent: HttpAgent;
+  /**
+   * The paths to request.
+   * @default []
+   */
   paths?: Path[] | Set<Path>;
-  blsVerify?: CreateCertificateOptions['blsVerify'];
+  /**
+   * Whether to disable the certificate freshness checks.
+   * @default false
+   */
+  disableCertificateTimeVerification?: boolean;
 };
 
 /**
- * Request information in the request_status state tree for a given canister.
+ * Requests information from a canister's `read_state` endpoint.
  * Can be used to request information about the canister's controllers, time, module hash, candid interface, and more.
- * @param {CanisterStatusOptions} options {@link CanisterStatusOptions}
- * @param {CanisterStatusOptions['canisterId']} options.canisterId {@link Principal}
- * @param {CanisterStatusOptions['agent']} options.agent {@link HttpAgent} optional authenticated agent to use to make the canister request. Useful for accessing private metadata under icp:private
- * @param {CanisterStatusOptions['paths']} options.paths {@link Path[]}
- * @returns {Status} object populated with data from the requested paths
+ * @param {CanisterStatusOptions} options The configuration for the canister status request.
+ * @see {@link CanisterStatusOptions} for detailed options.
+ * @returns {Promise<StatusMap>} A map populated with data from the requested paths. Each path is a key in the map, and the value is the data obtained from the certificate for that path.
  * @example
  * const status = await canisterStatus({
  *   paths: ['controllers', 'candid'],
@@ -138,12 +150,8 @@ export type CanisterStatusOptions = {
  *
  * const controllers = status.get('controllers');
  */
-export const request = async (options: {
-  canisterId: Principal;
-  agent: HttpAgent;
-  paths?: Path[] | Set<Path>;
-}): Promise<StatusMap> => {
-  const { agent, paths } = options;
+export const request = async (options: CanisterStatusOptions): Promise<StatusMap> => {
+  const { agent, paths, disableCertificateTimeVerification = false } = options;
   const canisterId = Principal.from(options.canisterId);
 
   const uniquePaths = [...new Set(paths)];
@@ -154,26 +162,27 @@ export const request = async (options: {
 
     return (async () => {
       try {
-        const response = await agent.readState(canisterId, {
-          paths: [encodedPath],
-        });
-        if (agent.rootKey == null) {
+        if (agent.rootKey === null) {
           throw ExternalError.fromCode(new MissingRootKeyErrorCode());
         }
 
-        const cert = await Certificate.create({
+        const rootKey = agent.rootKey;
+
+        const response = await agent.readState(canisterId, {
+          paths: [encodedPath],
+        });
+
+        const certificate = await Certificate.create({
           certificate: response.certificate,
-          rootKey: agent.rootKey,
-          canisterId: canisterId,
-          disableTimeVerification: true,
+          rootKey,
+          canisterId,
+          disableTimeVerification: disableCertificateTimeVerification,
+          agent,
         });
 
         const lookup = (cert: Certificate, path: Path) => {
           if (path === 'subnet') {
-            if (agent.rootKey == null) {
-              throw ExternalError.fromCode(new MissingRootKeyErrorCode());
-            }
-            const data = fetchNodeKeys(response.certificate, canisterId, agent.rootKey);
+            const data = fetchNodeKeys(response.certificate, canisterId, rootKey);
             return {
               path,
               data,
@@ -187,7 +196,7 @@ export const request = async (options: {
         };
 
         // must pass in the rootKey if we have no delegation
-        const { path, data } = lookup(cert, uniquePaths[index]);
+        const { path, data } = lookup(certificate, uniquePaths[index]);
         if (!data) {
           // Typically, the cert lookup will throw
           console.warn(`Expected to find result for path ${path}, but instead found nothing.`);
@@ -246,8 +255,12 @@ export const request = async (options: {
           }
         }
       } catch (error) {
-        // Break on signature verification errors
-        if (error instanceof AgentError && error.hasCode(CertificateVerificationErrorCode)) {
+        // Throw on certificate errors
+        if (
+          error instanceof AgentError &&
+          (error.hasCode(CertificateVerificationErrorCode) ||
+            error.hasCode(CertificateTimeErrorCode))
+        ) {
           throw error;
         }
         if (typeof path !== 'string' && 'key' in path && 'path' in path) {