+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - feat: adds the `hasSyncedTime` method to the `HttpAgent` class, which returns `true` if the time has been synced at least once with the IC network, `false` otherwise.
 - fix: use the effective canister id to delete the node keys from the local cache.
 - docs: add DFINITY Starlight theme to the docs
+- feat: adds the `disableCertificateTimeVerification` optional field to the `CanisterStatus.request` function argument, which allows you to control the `disableTimeVerification` option for the internal `Certificate.create` call.
 
 ## [3.1.0] - 2025-07-24